+++ src/server/auth/facebook/index.js
@@ -48,19 +48,19 @@ exports.init = function (passport, router, config) {
         }
     ));
 
-    router.get(config.loginURL, function*(next) {
-        this.session.redirect = `${config.successRedirect}?${this.request.querystring}`;
-        yield next;
+    router.get(config.loginURL, async (ctx, next) => {
+        ctx.session.redirect = `${config.successRedirect}?${ctx.request.querystring}`;
+        await next();
     }, passport.authenticate('facebook', {scope: ['email']}));
 
     router.get(config.callbackURL,
         passport.authenticate('facebook', {failureRedirect: config.failureRedirect}),
-        function*() {
+        async (ctx) => {
             // Successful authentication, redirect home.
-            if (this.session.redirect) {
-                this.response.redirect(this.session.redirect);
+            if (ctx.session.redirect) {
+                ctx.response.redirect(ctx.session.redirect);
             } else {
-                this.response.redirect(config.successRedirect);
+                ctx.response.redirect(config.successRedirect);
             }
         });
 };