+++ src/server/auth/google/index.js
@@ -55,9 +55,9 @@ exports.init = function (passport, router, config, mainConfig) {
         }
     ));
 
-    router.get('/login/google/popup', function * () {
-        this.session.popup = true;
-        this.redirect('/auth/login/google');
+    router.get('/login/google/popup', (ctx) => {
+        ctx.session.popup = true;
+        ctx.redirect('/auth/login/google');
     });
 
     router.get('/login/google', passport.authenticate('google', {scope: ['https://www.googleapis.com/auth/userinfo.email']}));