+++ test/basic.js
@@ -3,6 +3,7 @@
 const Couch = require('..');
 const nanoPromise = require('../src/util/nanoPromise');
 const assert = require('assert');
+const entryUnicity = require('./data/entryUnicity');
 
 process.on('unhandledRejection', function (err) {
     throw err;
@@ -35,24 +36,22 @@ describe('basic initialization tests', function () {
 });
 
 describe('basic initialization with custom design docs', function () {
-    let couch;
-    it('should load the design doc files at initialization', function () {
-        couch = Couch.get('test3');
-        return couch._initPromise.then(function () {
-            var app = nanoPromise.getDocument(couch._db, '_design/app')
-                .then(app => {
-                    assert.notEqual(app, null);
-                    assert.ok(app.views.test);
-                    assert.ok(app.filters.abc);
-                });
-            var custom = nanoPromise.getDocument(couch._db, '_design/custom')
-                .then(custom => {
-                    assert.notEqual(custom, null);
-                    assert.ok(custom.views.testCustom);
-                });
+    beforeEach(entryUnicity);
 
-            return Promise.all([app, custom]);
-        });
+    it('should load the design doc files at initialization', function () {
+        const app = nanoPromise.getDocument(couch._db, '_design/app')
+            .then(app => {
+                assert.notEqual(app, null);
+                assert.ok(app.views.test);
+                assert.ok(app.filters.abc);
+            });
+        const custom = nanoPromise.getDocument(couch._db, '_design/custom')
+            .then(custom => {
+                assert.notEqual(custom, null);
+                assert.ok(custom.views.testCustom);
+            });
+
+        return Promise.all([app, custom]);
     });
 
     it('should query a custom design document', function () {